test(web): rename console home e2e helper
